refactor(tests): split TransactionDraftTests (1994→≤293) (#272, #266, #267)
Clear the ERROR-level file_length violation (1994 > 1000) on
MoolahTests/Shared/TransactionDraftTests.swift by splitting along the
existing `// MARK:` seams into ten @Suite-per-file files, each well
under the 400-line warning threshold (largest is 293). Shared helpers
(`makeExpenseDraft`, `makeAccounts`, `makeAccount`, the account/instrument
constants) live in a new `MoolahTests/Support/TransactionDraftTestSupport.swift`.

Also eliminates the baselined violations that would otherwise re-fire
in the new paths:
  - type_body_length: the 1687-line struct is now ten structs ≤293 body-lines.
  - 35 × `tx` identifier_name → `transaction` (domain-descriptive at call site).
  - 69 × force_unwrapping → `try #require(...)`. The main patterns:
      draft.toTransaction(...)!   → try #require(draft.toTransaction(...))
      Decimal(string: "...")!     → try #require(Decimal(string: "..."))
                                    hoisted to a named local
      draft.legDrafts.last!       → try #require(draft.legDrafts.last)
    Tests were made `throws` where needed.

Stacked on #352 until it lands; retarget to main then.
